requirements.txt
alembic==1.17.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
dnspython==2.8.0
dominate==2.9.1
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pycparser==2.23
SQLAlchemy==2.0.44
typing_extensions==4.15.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
